# services/commerce/api/openapi.yaml
openapi: 3.0.3
info: {title: Commerce API, version: 1.0.0}
paths:
...
  /internal/operational-holds/{id}/convert:
    post:
      operationId: convertOperationalHold
      summary: Staff conversion of an operational hold into a purchasable reservation
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OperationalConversionCreate'}}}}
      responses:
        '200': {description: Replayed conversion, content: {application/json: {schema: {$ref: '#/components/schemas/OperationalConversion'}}}}
        '201': {description: Converted, content: {application/json: {schema: {$ref: '#/components/schemas/OperationalConversion'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
        '502': {$ref: '#/components/responses/Error'}
  /internal/group-reservations/{id}/draw-down:
    post:
      operationId: drawDownGroupReservation
      summary: Staff draw-down of a group/agency reservation into a purchasable reservation
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OperationalConversionCreate'}}}}
      responses:
        '200': {description: Replayed draw-down, content: {application/json: {schema: {$ref: '#/components/schemas/OperationalConversion'}}}}
        '201': {description: Drawn down, content: {application/json: {schema: {$ref: '#/components/schemas/OperationalConversion'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
        '502': {$ref: '#/components/responses/Error'}
...
components:
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 200}}
...
  responses:
    Error: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {type: string}
        code: {type: string, enum: [seat_taken, orphaned_seats], description: 'Machine-readable refusal reason. `seat_taken`: a seated reservation lost seats to a competing claimant (TKT-173). `orphaned_seats`: the selection would strand free seats with no free neighbour (ADR-041, TKT-182).'}
        seat_identities:
          type: array
          description: >
            With `seat_taken`: the requested seats another buyer already holds,
            sorted — a SUBSET of the request, forwarded verbatim from the inventory
            transaction that arbitrated. With `orphaned_seats`: the FREE seats the
            selection would strand — seats the buyer did NOT request, so the subset
            rule deliberately does not apply to them and a picker must keep them
            selectable, since adding one is the repair. Never synthesised by
            commerce: a refusal without a usable identity list is a 502, not a guess.
          items: {type: string, minLength: 1, maxLength: 200}
...
    OperationalConversionCreate:
      type: object
      additionalProperties: false
      required: [organizer_id, ticket_type_id, quantity, actor, reason]
      properties:
        organizer_id: {type: string, format: uuid}
        ticket_type_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, maximum: 50}
        actor: {type: string, minLength: 1, maxLength: 200}
        reason: {type: string, minLength: 1, maxLength: 500}
    OperationalConversion:
      type: object
      additionalProperties: false
      required: [reservation_id, hold_id, buyer_id, amount, currency, expires_at, server_time, source_remaining]
      properties:
        reservation_id: {type: string, format: uuid}
        hold_id: {type: string, format: uuid}
        buyer_id: {type: string, format: uuid}
        amount: {type: integer, format: int64}
        currency: {type: string}
        expires_at: {type: string, format: date-time}
        server_time: {type: string, format: date-time}
        source_remaining: {type: integer}
...